requirements.txt
face-recognition
pillow